Automating Input Fields and a click button
I created an automated click that takes you from the homepage to an
application form page then it check if the second page is been open an
run a JUnit test to prove the test. In the application form page it
runs automated input in the textbook fields present in the page.